backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
boto3==1.21.46
botocore==1.24.46
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.15
django-common-helpers==0.9.2
django-cors-headers==3.10.0
django-cron==0.5.1
django-debug-toolbar==3.4.0
django-environ==0.9.0
django-import-export==2.6.1
django-letsencrypt==4.0.0
django-test-without-migrations==0.6
django-tinymce==3.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
MarkupPy==1.14
maxminddb==2.2.0
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.2.0
pyotp==2.6.0
python-dateutil==2.8.1
python-http-client==3.3.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.2
sendgrid==6.9.7
six==1.15.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.61.0
tablib==3.0.0
urllib3==1.26.7
virtualenv==20.4.7
xlrd==2.0.1
xlwt==1.3.0
